arreglos, unit test, bug semiarreglado rubros

+++ db/migrate/20170409003850_create_rubro_service_details.rb
@@ -3,9 +3,8 @@ class CreateRubroServiceDetails < ActiveRecord::Migration[5.0]
     create_table :rubro_service_details do |t|
       t.references :rubro, foreign_key: true
       t.references :service, foreign_key: true
-      t.integer :quantity
+      t.float :quantity
       t.integer :subtotal
-
       t.timestamps
     end
   end
